common organization for multi apps
