+++ README.md
@@ -16,8 +16,11 @@ DOCKER_BUILDKIT=1 docker build -t <tag-name> .
 
 Examples:
 ```shell
-DOCKER_BUILDKIT=1 docker build -t ghcr.io/catalystneuro/dendro_si_kilosort25:latest .
+DOCKER_BUILDKIT=1 docker build -f si_kilosort25/Dockerfile -t ghcr.io/catalystneuro/dendro_si_kilosort25:latest .
 docker push ghcr.io/catalystneuro/dendro_si_kilosort25:latest
+
+DOCKER_BUILDKIT=1 docker build -f si_kilosort3/Dockerfile -t ghcr.io/catalystneuro/dendro_si_kilosort3:latest .
+docker push ghcr.io/catalystneuro/dendro_si_kilosort3:latest
 ```
 
 ## Test locally
