+++ common/processor_pipeline.py
@@ -1,0 +1,101 @@
+from spikeinterface_pipelines import pipeline as si_pipeline
+from spikeinterface.extractors import NwbRecordingExtractor
+import os
+import pynwb
+import h5py
+import logging
+
+from models import PipelineContext
+from nwb_utils import create_sorting_out_nwb_file
+
+
+logging.basicConfig(level=logging.INFO)
+logger = logging.getLogger(__name__)
+
+
+def run_pipeline(context: PipelineContext):
+    """
+    Runs the spikeinterface pipeline.
+
+    Args:
+        context (PipelineContext): Pipeline context model.
+    """
+    # Create SI recording from InputFile
+    logger.info('Opening remote input file')
+    download = not context.lazy_read_input
+    ff = context.input.get_file(download=download)
+
+    logger.info('Creating input recording')
+    recording = NwbRecordingExtractor(
+        file=ff,
+        electrical_series_location=context.recording_context.electrical_series_path,
+        # file_path=context.input.get_url(),
+        # stream_mode="remfile"
+    )
+
+    if context.stub_test:
+        logger.info('Running in stub test mode')
+        n_frames = int(min(300_000, recording.get_num_frames()))
+        recording = recording.frame_slice(start_frame=0, end_frame=n_frames)
+
+    logger.info(recording)
+
+    # Run pipeline
+    job_kwargs = {
+        'n_jobs': -1,
+        'chunk_duration': '1s',
+        'progress_bar': False
+    }
+
+    run_preprocessing = context.run_preprocessing
+    preprocessing_params = context.preprocessing_context.model_dump()
+    motion_correction_preset = preprocessing_params['motion_correction']['preset']
+    nonrigid_accurate_kwargs = preprocessing_params['motion_correction'].pop('motion_kwargs_nonrigid_accurate')
+    rigid_fast_kwargs = preprocessing_params['motion_correction'].pop('motion_kwargs_rigid_fast')
+    kilosort_like_kwargs = preprocessing_params['motion_correction'].pop('motion_kwargs_kilosort_like')
+    if motion_correction_preset == 'nonrigid_accurate':
+        preprocessing_params['motion_correction']['motion_kwargs'] = nonrigid_accurate_kwargs
+    elif motion_correction_preset == 'rigid_fast':
+        preprocessing_params['motion_correction']['motion_kwargs'] = rigid_fast_kwargs
+    elif motion_correction_preset == 'kilosort_like':
+        preprocessing_params['motion_correction']['motion_kwargs'] = kilosort_like_kwargs
+
+    run_spikesorting = context.run_spikesorting
+    spikesorting_params = context.spikesorting_context.model_dump()
+
+    run_postprocessing = context.run_postprocessing
+    # postprocessing_params = context.postprocessing_context.model_dump()
+
+    logger.info('Running pipeline')
+    _, sorting, _ = si_pipeline.run_pipeline(
+        recording=recording,
+        scratch_folder="./scratch/",
+        results_folder="./results/",
+        job_kwargs=job_kwargs,
+        run_preprocessing=run_preprocessing,
+        preprocessing_params=preprocessing_params,
+        run_spikesorting=run_spikesorting,
+        spikesorting_params=spikesorting_params,
+        run_postprocessing=run_postprocessing,
+        # postprocessing_params=postprocessing_params,
+    )
+
+    # Upload output file
+    if sorting:
+        logger.info('Writing output NWB file')
+        h5_file = h5py.File(ff, 'r')
+        with pynwb.NWBHDF5IO(file=h5_file, mode='r', load_namespaces=True) as io:
+            nwbfile_rec = io.read()
+
+            if not os.path.exists('output'):
+                os.mkdir('output')
+            sorting_out_fname = 'output/sorting.nwb'
+
+            create_sorting_out_nwb_file(
+                nwbfile_original=nwbfile_rec,
+                sorting=sorting,
+                sorting_out_fname=sorting_out_fname
+            )
+
+        logger.info('Uploading output NWB file')
+        context.output.upload(sorting_out_fname)
