+++ si_kilosort3/Dockerfile
@@ -6,12 +6,24 @@ RUN git clone https://github.com/flatironinstitute/dendro.git && \
     cd dendro/python && \
     pip install -e .
 
+# Install spikeinterface-pipelines from source, for now
+RUN git clone https://github.com/SpikeInterface/spikeinterface_pipelines.git && \
+    cd spikeinterface_pipelines && \
+    git checkout dev && \
+    pip install -e .
+
+# Install spikeinterface from source, for now
+RUN git clone https://github.com/SpikeInterface/spikeinterface.git && \
+    cd spikeinterface && \
+    # git checkout dev && \
+    pip install -e .[full]
+
 # Copy files into the container
 WORKDIR /app
-COPY *.py ./
-
-# Install requirements
+COPY /common /app/common
+COPY si_kilosort3/*.py /app/
+COPY si_kilosort3/requirements.txt /app/requirements.txt
 RUN pip install -r requirements.txt
 
-# Set the entrypoint
-ENTRYPOINT ["python", "main.py"]
+# Give execution permission to /app/main.py
+RUN chmod +x /app/main.py
