+++ si_kilosort3/requirements.txt
@@ -1,1 +1,4 @@
-spikeinterface[full]==0.98.2
+pynwb
+# spikeinterface[full]==0.99.1
+# spikeinterface_pipelines
+# dendro
